description of flipsign(x,y)
I changed flipsign(x,y) description from "a value with the magnitude of `x` and the sign of `x*y`" to "a value with the magnitude of `x` and the sign of `x.*y`", i.e., from  `x*y` to `x.*y`. This way delivers readers a better understanding on this function esp. when x and y are arrays. thank you.